+++ lib/ramp_flutter.dart
@@ -1,12 +1,12 @@
 import 'package:flutter/widgets.dart';
-import 'package:ramp_flutter/configuration.dart';
-import 'package:ramp_flutter/host_event.dart';
+import 'package:ramp_flutter/src/configuration.dart';
+import 'package:ramp_flutter/src/host_event.dart';
 import 'package:ramp_flutter/src/ramp_webview.dart';
-import 'package:ramp_flutter/widget_event.dart';
+import 'package:ramp_flutter/src/widget_event.dart';
 
-export 'package:ramp_flutter/configuration.dart';
-export 'package:ramp_flutter/host_event.dart';
-export 'package:ramp_flutter/widget_event.dart';
+export 'package:ramp_flutter/src/configuration.dart';
+export 'package:ramp_flutter/src/host_event.dart';
+export 'package:ramp_flutter/src/widget_event.dart';
 
 class RampFlutter {
   RampFlutter(Uri widgetUrl) : _webView = RampWebView(widgetUrl);